Convert to worker scoped fixtures
