+++ test/integration/editor/new.spec.ts
@@ -1,0 +1,45 @@
+import { test, expect } from '../../playwright/localTest';
+import { ToolpadEditor } from '../../models/ToolpadEditor';
+
+test.use({
+  localAppConfig: {
+    cmd: 'dev',
+  },
+});
+
+test('can place new components from catalog', async ({ page }) => {
+  const editorModel = new ToolpadEditor(page);
+
+  await editorModel.goto();
+
+  await editorModel.pageRoot.waitFor();
+
+  const canvasInputLocator = editorModel.appCanvas.locator('input');
+
+  await expect(canvasInputLocator).toHaveCount(0);
+
+  const TEXT_FIELD_COMPONENT_DISPLAY_NAME = 'Text field';
+
+  // Drag in a first component
+
+  await editorModel.dragNewComponentToAppCanvas(TEXT_FIELD_COMPONENT_DISPLAY_NAME);
+
+  await expect(canvasInputLocator).toHaveCount(1);
+  await expect(canvasInputLocator).toBeVisible();
+  expect(await page.getByLabel('name').inputValue()).toBe('textField');
+
+  // Drag in a second component
+
+  await editorModel.dragNewComponentToAppCanvas(TEXT_FIELD_COMPONENT_DISPLAY_NAME);
+
+  await expect(canvasInputLocator).toHaveCount(2);
+});
+
+test('can create new component', async ({ page }) => {
+  const editorModel = new ToolpadEditor(page);
+
+  await editorModel.goto();
+
+  await editorModel.createPage('somePage');
+  await editorModel.createComponent('someComponent');
+});
